backend/requirements.txt
Flask==2.2.2
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
Flask-Migrate==3.1.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
psycopg2-binary==2.9.3
